requirements_flask.txt
Flask==2.3.3
flask[async]==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
debugpy # Required for debugging.
pyyaml
python-dotenv
google-api-python-client
pyyaml
pydash
# rich

#cms
firebase_admin

#townscript
pandas  #for townscript
gspread

pytest
